+++ paq8pxd.cpp
@@ -21,7 +21,7 @@
 */
 
 #define PROGNAME "paq8pxd111"  // Please change this if you change the program.
-#define SIMD_GET_SSE  //uncomment to use SSE2 in ContexMap
+
 //#define MT            //uncomment for multithreading, compression only. Handled by CMake and gcc when -DMT is passed.
 #ifndef DISABLE_SM
 //#define SM              // For faster statemap
@@ -118,6 +118,10 @@ inline int max(int a, int b) {return a<b?b:a;}
 #include "prt/statemap.hpp"
 #include "prt/statetable.hpp"
 #include "prt/APM.hpp"
+#include "prt/indirect.hpp"
+#include "prt/indirectcontext.hpp"
+#include "prt/ols.hpp"
+#include "prt/largestationarymap.hpp"
 
 /////////////////////// Global context /////////////////////////
 U8 level=DEFAULT_OPTION;  // Compression level 0 to 15
@@ -494,387 +498,8 @@ public:
 };
 
 
-#pragma pack(push,1)
-template<typename T>
-struct HashElement {
-  uint16_t checksum;
-  T value;
-};
-#pragma pack(pop)
-
-template<typename T, int ElementsInBucket>
-class Bucket16 {
-private:
-  HashElement<T> elements[ElementsInBucket];
-public:
-
-  void reset() {
-    for (size_t i = 0; i < ElementsInBucket; i++)
-      elements[i] = {};
-  }
-
-  void stat(uint64_t& used, uint64_t& empty) {
-    for (size_t i = 0; i < ElementsInBucket; i++)
-      if (elements[i].checksum == 0)
-        empty++;
-      else
-        used++;
-  }
-
-  T* find(uint16_t checksum, Random1* rnd) {
-
-    checksum += checksum == 0; //don't allow 0 checksums (0 checksums are used for empty slots)
-
-    if (elements[0].checksum == checksum) //there is a high chance that we'll find it in the first slot, so go for it
-      return &elements[0].value;
-
-    for (size_t i = 1; i < ElementsInBucket; ++i) {
-      if (elements[i].checksum == checksum) { // found matching checksum
-        T value = elements[i].value;
-        //shift elements down
-        memmove(&elements[1], &elements[0], i * sizeof(HashElement<T>));
-        //move element to front (re-create)
-        elements[0].checksum = checksum;
-        elements[0].value = value;
-        return &elements[0].value;
-      }
-      if (elements[i].checksum == 0) { // found empty slot
-        //shift elements down (free the first slot for the new element)
-        memmove(&elements[1], &elements[0], i * sizeof(HashElement<T>)); // i==0 is OK
-        goto create_element;
-      }
-    }
-
-    //no match and no empty slot -> overwrite an existing element with the new (empty) one
-
-    //Replacement strategy:
-    // - In case the hash elements represent an indirect context (we have content but no priority): overwrite the last used element
-    // - In case the hash elements are counts/statistics: overwrite one of the the last used elements having low priority.
-    //
-    //   The priority of an element is established by the counts it represents: large counts usually mean that the element is worthy to keep.
-    //   When an element is accessed recently it is probably worthy to keep as it represents recent statistics,
-    //   when an element is accessed long ago it is probably obolete and may be overwritten.
-    //   In order to keep track of the order the elements were accessed we move the most recently accessed element to the front: thus
-    //   recently acessed elements are usually somewhere in the front of the bucket, rarely accessed element slowly move to the back. The
-    //   last elements in the bucket are cantidates for eviction.
-    //
-    //   Overwriting the lowest priority element (with no regard of its position) is in favor of homogenous (semi-stationary) files.
-    //   Overwriting the last element (with no regard of its priority) is in favor of mixed-content files.
-    //   In order to favor both, we are using a probabilistic replacement startegy where most recently accessed and/or higher priority
-    //   elements have higher chance to stay in the bucket = rarely accessed and/or low priority elements have higher chance to be 
-    //   evicted (overwritten).
-    //
-    //   The the 2 most recently accessed elements are always protected from overwriting.
-    {
-      size_t minElementIdx = ElementsInBucket - 1;
-     // if constexpr (std::is_same<T, HashElementForContextMap>::value || std::is_same<T, HashElementForStationaryMap>::value)
-     {
-        uint32_t RND = rnd->operator()(32);
-        if ((RND & 63) >= 1) {
-          RND >>= 6;
-          uint32_t minPrio = elements[ElementsInBucket - 1].value.prio();
-          uint32_t thisPrio = elements[ElementsInBucket - 2].value.prio();
-          if (thisPrio < minPrio) {
-            minPrio = thisPrio;
-            minElementIdx = ElementsInBucket - 2;
-          }
-          if ((RND & 63) >= 4) {
-            RND >>= 6;
-            thisPrio = elements[ElementsInBucket - 3].value.prio();
-            if (thisPrio < minPrio) {
-              minPrio = thisPrio;
-              minElementIdx = ElementsInBucket - 3;
-            }
-            if ((RND & 63) >= 8) {
-              RND >>= 6;
-              thisPrio = elements[ElementsInBucket - 4].value.prio();
-              if (thisPrio < minPrio) {
-                minPrio = thisPrio;
-                minElementIdx = ElementsInBucket - 4;
-              }
-              if ((RND & 63) >= 16) {
-                //RND >>= 6; //not necessary
-                thisPrio = elements[ElementsInBucket - 5].value.prio();
-                if (thisPrio < minPrio) {
-                  minPrio = thisPrio;
-                  minElementIdx = ElementsInBucket - 5;
-                }
-              }
-            }
-          }
-        }
-      }
-
-      //shift elements down (make room for the new element in the first slot)
-      //at the same time owerwrite the element at position "minElementIdx"
-      memmove(&elements[1], &elements[0], minElementIdx * sizeof(HashElement<T>));
-    }
-
-  create_element:
-    elements[0].checksum = checksum;
-    elements[0].value = {};
-    return &elements[0].value;
-  }
-};
 
 
-template<typename T, int ElementsInBucket>
-class B16 {
-private:
-  HashElement<T> elements[ElementsInBucket];
-public:
-  void reset() {
-    for (size_t i = 0; i < ElementsInBucket; i++)
-      elements[i] = {};
-  }
-
-  void stat(uint64_t& used, uint64_t& empty) {
-    for (size_t i = 0; i < ElementsInBucket; i++)
-      if (elements[i].checksum == 0)
-        empty++;
-      else
-        used++;
-  }
-
-  T* find(uint16_t checksum, Random1* rnd) {
-    checksum += checksum == 0; //don't allow 0 checksums (0 checksums are used for empty slots)
-    if (elements[0].checksum == checksum) //there is a high chance that we'll find it in the first slot, so go for it
-      return &elements[0].value;
-
-    for (size_t i = 1; i < ElementsInBucket; ++i) {
-      if (elements[i].checksum == checksum) { // found matching checksum
-        T value = elements[i].value;
-        //shift elements down
-        memmove(&elements[1], &elements[0], i * sizeof(HashElement<T>));
-        //move element to front (re-create)
-        elements[0].checksum = checksum;
-        elements[0].value = value;
-        return &elements[0].value;
-      }
-      if (elements[i].checksum == 0) { // found empty slot
-        //shift elements down (free the first slot for the new element)
-        memmove(&elements[1], &elements[0], i * sizeof(HashElement<T>)); // i==0 is OK
-        elements[0].checksum = checksum;
-        elements[0].value = {};
-        return &elements[0].value;
-      }
-    }
-    size_t minElementIdx = ElementsInBucket - 1;
-    //shift elements down (make room for the new element in the first slot)
-    //at the same time owerwrite the element at position "minElementIdx"
-    memmove(&elements[1], &elements[0], minElementIdx * sizeof(HashElement<T>));
-    elements[0].checksum = checksum;
-    elements[0].value = {};
-    return &elements[0].value;
-  }
-};
-
-struct HashElementForStationaryMap { // sizeof(HashElemetForStationaryMap) = 4
-  uint32_t value;
-  // priority for hash replacement strategy
-  uint32_t prio() {
-    return (value >> 16) + (value & 0xffff); // to be tuned
-  }
-};
-/**
- * Map for modelling contexts of (nearly-)stationary data.
- * The context is a hash. For each bit modelled, the exact counts of 0s and 1s are stored.
- *
- */
-
-class LargeStationaryMap  {
-private:
-    Random1 rnd;
-    Array<Bucket16<HashElementForStationaryMap, 7>> data;
-    const uint32_t hashBits;
-    int scale, rate;
-    uint64_t context;
-    uint32_t *cp;
-public:
-    /**
-     * Construct using 2^hashBits * sizeof(Bucket16) bytes of memory for storing a maximum of 2^hashBits * ElementsInBucket
-     * That is:
-      *   hashBits    |       memory          |   maximum number of contexts
-      *       10      |  2^10 * 42 =   43 KB  |         7 K  (~12 bits)
-      *       11      |  2^11 * 42 =   86 KB  |        14 K  (~13 bits)
-      *       12      |  2^12 * 42 =  172 KB  |        28 K  (~14 bits)
-      *       13      |  2^13 * 42 =  344 KB  |        57 K  (~15 bits)
-      *       14      |  2^14 * 42 =  688 KB  |       114 K  (~16 bits)
-      *       15      |  2^15 * 42 =  1.3 MB  |       229 K  (~17 bits)
-      *       16      |  2^16 * 42 =  2.7 MB  |       458 K  (~18 bits)
-      *       17      |  2^17 * 42 = 11.0 MB  |       1.8 M  (~19 bits)
-      *       18      |  2^18 * 42 = 22.0 MB  |       3.6 M  (~20 bits)
-      *       19      |  2^19 * 42 = 44.0 MB  |       7.3 M  (~21 bits)
-      *       20      |  2^20 * 42 = 88.0 MB  |      14.6 M  (~22 bits)
-      *      ...              ...                         ...
-      * 
-     * @param scale
-     * @param rate use 16 near-stationary modelling (default), smaller values may be used for tuning adaptivity
-     */
-    LargeStationaryMap( const int hashBits, const int scale = 64, const int rate = 16);
-    /**
-     * ctx must be a hash
-     * @param ctx
-     */
-    void set(uint64_t ctx);
-    void setscale(int scale);
-    void reset();
-    void update(Mixer &m) ;
-    void mix(Mixer &m);
-};
-
-LargeStationaryMap::LargeStationaryMap( const int hashBits, const int scale, const int rate) :
-  
-  rnd(),
-  data((UINT64_C(1) << hashBits)),
-  hashBits(hashBits),
-  scale(scale),
-  rate(rate) {
-#ifdef VERBOSE
-  printf("Created LargeStationaryMap with hashBits = %d, %d, scale = %d, rate = %d\n", hashBits, scale, rate);
-#endif
-  assert(hashBits > 0);
-  assert(hashBits <= 24); // 24 is just a reasonable limit for memory use 
-  assert(9 <= rate && rate <= 16); // 9 is just a reasonable lower bound, 16 is a hard bound
-  reset();
-  set(0);
-  cp = &data[0].find(0, &rnd)->value;
-}
-
-void LargeStationaryMap::set(uint64_t ctx) {
-  context = ctx;
-}
-
-void LargeStationaryMap::setscale(int scale) {
-  this->scale = scale;
-}
-
-void LargeStationaryMap::reset() {
-  for( uint32_t i = 0; i < data.size(); ++i ) {
-    data[i].reset();
-  }
-}
-
-void LargeStationaryMap::update(Mixer &m) {
-  uint32_t n0, n1, value;
-  value = *cp;
-  n0 = value >> 16;
-  n1 = value & 0xffff;
-
-  n0 += 1 - m.x.y;
-  n1 += m.x.y;
-  int shift = (n0 | n1) >> rate; // shift: 0 or 1
-  n0 >>= shift;
-  n1 >>= shift;
-
-  *cp = n0 << 16 | n1;
-
-  context = hash(context, m.x.y);
-}
-uint16_t checksum16(const uint64_t hash, const int hashBits) {
-  constexpr int checksumBits = 16;
-  return static_cast<uint16_t>(hash >> (64 - hashBits - checksumBits)) & ((1U << checksumBits) - 1);
-}
-void LargeStationaryMap::mix(Mixer &m) {
-  update(m);
-  uint32_t n0, n1, value, sum;
-  int p1, st, bitIsUncertain;
-  
-  uint32_t hashkey = finalize64(context, hashBits);
-  uint16_t checksum = checksum16(context, hashBits);
-  cp = &data[hashkey].find(checksum, &rnd)->value;
-  value = *cp;
-  n0 = value >> 16;
-  n1 = value & 0xffff;
-
-  sum = n0 + n1;
-  p1 = ((n1 * 2 + 1) << 12) / (sum * 2 + 2);
-  st = (stretch(p1) * scale) >> 8;
-  m.add(st);
-  m.add(((p1 - 2048) * scale) >> 9);
-  bitIsUncertain = int(sum <= 1 || (n0 != 0 && n1 != 0));
-  m.add((bitIsUncertain - 1) & st); // when both counts are nonzero add(0) otherwise add(st)
-  //p0 = 4095 - p1;
-  //m.add((((p1 & (-!n0)) - (p0 & (-!n1))) * scale) >> 10);
-}
-class IndirectMap {
-  Array<U8> Data;
-  StateMap Map;
-  const int mask, maskbits, stride;
-  int Context, bCount, bTotal, B;
-  U8 *cp;
-  Random rnd;
-  int Limit;
-  int scale;
-public:
-  IndirectMap(int BitsOfContext, int InputBits = 8,int s=256,int l=1023): 
-  Data((1ull<<BitsOfContext)*((1ull<<InputBits)-1)), mask((1<<BitsOfContext)-1),
-   maskbits(BitsOfContext), stride((1<<InputBits)-1), Context(0),
-    bCount(0), bTotal(InputBits), B(0),scale(s),Limit(l) {
-    assert(InputBits>0 && InputBits<=8);
-    assert(BitsOfContext+InputBits<=24);
-    cp=&Data[0];
-  }
-  void set_direct(const U32 ctx) {
-    Context = (ctx&mask)*stride;
-    bCount=B=0;
-  }
-  void set(const U64 ctx) {
-    Context = (finalize64(ctx,maskbits)&mask)*stride;
-    bCount=B=0;
-  }
-  void mix(Mixer& m, const int Multiplier = 1, const int Divisor = 4, const U16 Limit = 1023) {
-    // update
-    //*cp = nex(*cp, m.x.y);
-    int ns=nex(*cp, m.x.y);
-      if (ns>=204 && rnd() << ((452-ns)>>3)) ns-=4;  // probabilistic increment
-      *cp=ns;
-    // predict
-    B+=(m.x.y && B>0);
-    cp=&Data[Context+B];
-    const U8 state = *cp;
-    const int p1 = Map.p(state,m.x.y, Limit);
-    m.add((stretch(p1)*Multiplier)/Divisor);
-    m.add(((p1-2048)*Multiplier)/(Divisor*2));
-    bCount++; B+=B+1;
-    if (bCount==bTotal)
-      bCount=B=0;
-  }
-  void mix1(Mixer& m) {
-    // update
-    //*cp = nex(*cp, m.x.y);
-    int ns=nex(*cp, m.x.y);
-      if (ns>=204 && rnd() << ((452-ns)>>3)) ns-=4;  // probabilistic increment
-      *cp=ns;
-    // predict
-    B+=(m.x.y && B>0);
-    cp=&Data[Context+B];
-    const U8 state = *cp;
-    const int p1 = Map.p1(state,m.x.y, Limit);
-    m.add((stretch(p1)*scale  )>> 8U);
-    m.add(((p1-2048)*scale  )>> 9U);
-    bCount++; B+=B+1;
-    if (bCount==bTotal)
-      bCount=B=0;
-  }
-  void mix2(Mixer& m) {
-    // update
-    //*cp = nex(*cp, m.x.y);
-    int ns=nex(*cp, m.x.y);
-      if (ns>=204 && rnd() << ((452-ns)>>3)) ns-=4;  // probabilistic increment
-      *cp=ns;
-    // predict
-    B+=(m.x.y && B>0);
-    cp=&Data[Context+B];
-    const U8 state = *cp;
-    const int p1 = Map.p2(state,m.x.y, Limit);
-    m.add((stretch(p1)*scale  )>> 8U);
-    m.add(((p1-2048)*scale  )>> 9U);
-    bCount++; B+=B+1;
-    if (bCount==bTotal)
-      bCount=B=0;
-  }
-};
 
 // Context map for large contexts.  Most modeling uses this type of context
 // map.  It includes a built in RunContextMap to predict the last byte seen
@@ -969,118 +594,18 @@ public:
     int inputs();
 };
 
-#if defined(SIMD_GET_SSE) 
-
-#define xmmbshl(x,n)  _mm_slli_si128(x,n) // xm <<= 8*n  -- BYTE shift left
-#define xmmbshr(x,n)  _mm_srli_si128(x,n) // xm >>= 8*n  -- BYTE shift right
-#define xmmshl64(x,n) _mm_slli_epi64(x,n) // xm.hi <<= n, xm.lo <<= n
-#define xmmshr64(x,n) _mm_srli_epi64(x,n) // xm.hi >>= n, xm.lo >>= n
-#define xmmand(a,b)   _mm_and_si128(a,b)
-#define xmmor(a,b)    _mm_or_si128(a,b)
-#define xmmxor(a,b)   _mm_xor_si128(a,b)
-#define xmmzero       _mm_setzero_si128()
-#ifdef _MSC_VER
-#include <intrin.h>
-U32 __inline clz(U32 value){ //asume newer version of vc
-    unsigned long leading_zero = 0;
-    if (_BitScanReverse( &leading_zero, value)) return 31-leading_zero;
-    else return 32;
-}
-  U32 __inline ctz(U32 x ){
-    unsigned long leading_zero = 0;
-   if (_BitScanForward(&leading_zero, x )) {
-       return  leading_zero;
-    }
-    else{
-         return 0; // Same remarks as above
-    }
-}
-
-#elif ((__GNUC__ >= 4) || ((__GNUC__ == 3) && (__GNUC_MINOR__ >= 4)))
-U32 __inline clz(U32 value){ 
-    return __builtin_clz(value);
-}
-U32 __inline ctz(U32 value){ 
-    return __builtin_ctz(value);
-}
-#endif
-#endif
 
 // Find or create hash element matching checksum ch
 inline U8* ContextMap::E::get(U16 ch) {
     
   if (chk[last&15]==ch) return &bh[last&15][0];
   int b=0xffff, bi=0;
-#if defined(SIMD_GET_SSE)   
-  const XMM xmmch=_mm_set1_epi16 (short(ch)); //fill 8 ch values
-//load 8 values, discard last one as only 7 are needed.
-//reverse order and compare 7 chk values to ch
-//get mask is set get first index and return value  
-  XMM tmp=_mm_load_si128 ((XMM *) &chk[0]); //load 8 values (8th will be discarded)
-#if defined(__SSSE3__)
-#include <immintrin.h>
- // const XMM vm=;// initialise vector mask 
-  tmp=_mm_shuffle_epi8(tmp,_mm_setr_epi8(14,15,12,13,10,11,8,9,6,7,4,5,2,3,0,1));   
-#elif   defined(__SSE2__) 
-  tmp=_mm_shufflelo_epi16(tmp,0x1B); //swap order for mask  (0,1,2,3)
-  tmp=_mm_shufflehi_epi16(tmp,0x1B);                      //(0,1,2,3)
-  tmp=_mm_shuffle_epi32(tmp,0x4E);                        //(1,0,3,2)   
-     
-#endif
-  tmp=_mm_cmpeq_epi16 (tmp,xmmch); //compare ch values
-  tmp=_mm_packs_epi16(tmp,xmmzero); //pack result
-  U32 t=(_mm_movemask_epi8(tmp))>>1; //get mask of comparsion, bit is set if eq, discard 8th bit
-  U32 a;    //index into bh or 7 if not found
-  if(t){
-      a=(clz(t)-1)&7;
-      return last=last<<4|a, (U8*)&bh[a][0];
-  }
-
- XMM   lastl=_mm_set1_epi8((last&15));
- XMM   lasth=_mm_set1_epi8((last>>4));
- XMM   one1  =_mm_set1_epi8(1);
- XMM   vm=_mm_setr_epi8(0,1,2,3,4,5,6,7,0,1,2,3,4,5,6,7);
-
- XMM   lastx=_mm_unpacklo_epi64(lastl,lasth); //last&15 last>>4
- XMM   eq0  =_mm_cmpeq_epi8 (lastx,vm); //compare   values
-
- eq0=_mm_or_si128(eq0,_mm_srli_si128 (eq0, 8));    //or low values with high
-
- lastx = _mm_and_si128(one1, eq0);                //set to 1 if eq
- XMM sum1 = _mm_sad_epu8(lastx,xmmzero);        //cout values, abs(a0 - b0) + abs(a1 - b1) .... up to b8
- const U32 pcount=_mm_cvtsi128_si32(sum1); //population count
-/*for (int i=0; i<7; ++i) {
-   bh[i][0]=i+1;
-    
-  }*/
- U32 t0=(~_mm_movemask_epi8(eq0));
-for (int i=pcount; i<7; ++i) {
-    int bitt =ctz(t0);     //get index 
-//#if ((__GNUC__ >= 4) || ((__GNUC__ == 3) && (__GNUC_MINOR__ >= 4)))    
-//asm("btr %1,%0" : "+r"(t0) : "r"(bitt)); // clear bit set and test again https://gcc.gnu.org/bugzilla/show_bug.cgi?id=47769
-//#else
-    t0 &= ~(1 << bitt); // clear bit set and test again
-//#endif 
-   int pri=bh[bitt][0];
-    if (pri<b  ) b=pri, bi=bitt;
-
-
-  }  
-  /*
-  //uncomment above SIMD version and comment out code below to use full SIMD (SSE2) version
-  for (int i=0; i<7; ++i) {
-    int pri=bh[i][0];
-    if (pri<b && (last&15)!=i && (last>>4)!=i) b=pri, bi=i;
-  }*/
-  return last=0xf0|bi, chk[bi]=ch, (U8*)memset(&bh[bi][0], 0, 7);
-#else
   for (int i=0; i<7; ++i) {
     if (chk[i]==ch) return last=last<<4|i, (U8*)&bh[i][0];
     int pri=bh[i][0];
     if (pri<b && (last&15)!=i && last>>4!=i) b=pri, bi=i;
   }
   return last=0xf0|bi, chk[bi]=ch, (U8*)memset(&bh[bi][0], 0, 7);
-#endif
 }
 
 // Construct using m bytes of memory for c contexts(c+7)&-8
@@ -1586,230 +1111,8 @@ public:
   int inputs() {return inputCount;}
 };
 
-///////////////// Ordinary Least Squares predictor /////////////////
 
-template <typename F, typename T, const bool hasZeroMean = true>
-class OLS {
-  static constexpr F ftol = 1E-8;
-  static constexpr F sub = F(int64_t(!hasZeroMean)<<(8*sizeof(T)-1));
-private:
-  int n, kmax, km, index;
-  F lambda, nu;
-  F *x, *w, *b;
-  F **mCovariance, **mCholesky;
-  int Factor() {
-    // copy the matrix
-    for (int i=0; i<n; i++)
-      for (int j=0; j<n; j++)
-        mCholesky[i][j] = mCovariance[i][j];
 
-    for (int i=0; i<n; i++)
-      mCholesky[i][i] += nu;
-    for (int i=0; i<n; i++) {
-      for (int j=0; j<i; j++) {
-        F sum = mCholesky[i][j];
-        for (int k=0; k<j; k++)
-          sum -= (mCholesky[i][k] * mCholesky[j][k]);
-        mCholesky[i][j] = sum / mCholesky[j][j];
-      }
-      F sum = mCholesky[i][i];
-      for (int k=0; k<i; k++)
-        sum -= (mCholesky[i][k] * mCholesky[i][k]);
-      if (sum>ftol)
-        mCholesky[i][i] = sqrt(sum);
-      else
-        return 1;
-    }
-    return 0;
-  }
-
-  void Solve() {
-    for (int i=0; i<n; i++) {
-      F sum = b[i];
-      for (int j=0; j<i; j++)
-        sum -= (mCholesky[i][j] * w[j]);
-      w[i] = sum / mCholesky[i][i];
-    }
-    for (int i=n-1; i>=0; i--) {
-      F sum = w[i];
-      for (int j=i+1; j<n; j++)
-        sum -= (mCholesky[j][i] * w[j]);
-      w[i] = sum / mCholesky[i][i];
-    }
-  }
-public:
-  OLS(int n, int kmax=1, F lambda=0.998, F nu=0.001) : n(n), kmax(kmax), lambda(lambda), nu(nu) {
-    km = index = 0;
-    x = new F[n], w = new F[n], b = new F[n];
-    mCovariance = new F*[n], mCholesky = new F*[n];
-    for (int i=0; i<n; i++) {
-      x[i] = w[i] = b[i] = 0.;
-      mCovariance[i] = new F[n], mCholesky[i] = new F[n];
-      for (int j=0; j<n; j++)
-        mCovariance[i][j] = mCholesky[i][j] = 0.;
-    }
-  }
-  ~OLS() {
-    delete x, delete w, delete b;
-    for (int i=0; i<n; i++) {
-      delete mCovariance[i];
-      delete mCholesky[i];
-    }
-    delete[] mCovariance, delete[] mCholesky;
-  }
-  void Add(const T val) {
-    if (index<n)
-      x[index++] = F(val)-sub;
-  }
-  void AddFloat(const F val) {
-    if (index<n)
-      x[index++] = val-sub;
-  }
-  F Predict(const T **p) {
-    F sum = 0.;
-    for (int i=0; i<n; i++)
-      sum += w[i] * (x[i] = F(*p[i])-sub);
-    return sum+sub;
-  }
-  F Predict() {
-    assert(index==n);
-    index = 0;
-    F sum = 0.;
-    for (int i=0; i<n; i++)
-      sum += w[i] * x[i];
-    return sum+sub;
-  }
-  void Update(const T val) {
-    for (int j=0; j<n; j++)
-      for (int i=0; i<n; i++)
-        mCovariance[j][i] = lambda * mCovariance[j][i] + (1.0 - lambda) * (x[j] * x[i]);
-    for (int i=0; i<n; i++)
-      b[i] = lambda * b[i] + (1.0 - lambda) * (x[i] * (F(val)-sub));
-    km++;
-    if (km>=kmax) {
-      if (!Factor()) Solve();
-      km = 0;
-    }
-  }
-};
-////////////////////////////// Indirect Context //////////////////////////////
-template <typename T>
-class IndirectContext {
-private:
-  Array<T> data;
-  T* ctx;
-  U32 ctxMask, inputMask, inputBits;
-public:
-  IndirectContext(const int BitsPerContext, const int InputBits = 8) :
-    data(1ull<<BitsPerContext),
-    ctx(&data[0]),
-    ctxMask((1ul<<BitsPerContext)-1),
-    inputMask((1ul<<InputBits)-1),
-    inputBits(InputBits)
-  {
-    assert(BitsPerContext>0 && BitsPerContext<=20);
-    assert(InputBits>0 && InputBits<=8);
-  }
-  void operator+=(const U32 i) {
-    assert(i<=inputMask);
-    (*ctx)<<=inputBits;
-    (*ctx)|=i;
-  }
-  void operator=(const U32 i) {
-    ctx = &data[i&ctxMask];
-  }
-  T& operator()(void) {
-    return *ctx;
-  }
-};
-
-template<typename T>
-class IndirectContext1 {
-private:
-    Array<T> data;
-    T *ctx;
-    const uint32_t ctxMask, inputMask, inputBits, contextBits;
-
-public:
-    IndirectContext1(const int bitsPerContext, const int inputBits, const int contextBits = sizeof(T)*8) :
-      data(UINT64_C(1) << bitsPerContext), ctx(&data[0]),
-      ctxMask((UINT32_C(1) << bitsPerContext) - 1), 
-      inputMask((UINT32_C(1) << inputBits) - 1), 
-      inputBits(inputBits),
-      contextBits(contextBits) {
-#ifdef VERBOSE
-      printf("Created IndirectContext with bitsPerContext = %d, inputBits = %d\n", bitsPerContext, inputBits);
-#endif
-      assert(bitsPerContext > 0 && bitsPerContext <= 20);
-      assert(inputBits > 0 && inputBits <= 8);
-      assert(contextBits <= sizeof(T)*8);
-      if (contextBits < sizeof(T) * 8) // need for leading bit -> include it
-        reset(); 
-    }
-
-    void reset() {
-      for (uint64_t i = 0; i < data.size(); ++i) {
-        data[i] = contextBits < sizeof(T) * 8 ? 1 : 0; // 1: leading bit to indicate the number of used bits
-      }
-    }
-
-    void operator+=(const uint32_t i) {
-      assert(i <= inputMask);
-      // note: when the context is fully mature, we need to keep the leading bit in front of the contextbits as the MSB
-      T leadingBit = (*ctx) & (1 << contextBits); 
-      (*ctx) <<= inputBits;
-      (*ctx) |= i | leadingBit;
-      (*ctx) &= (1 << (contextBits + 1)) - 1;
-    };
-
-    void operator=(const uint32_t i) {
-      ctx = &data[i & ctxMask];
-    }
-
-    auto operator()() -> T & {
-      return *ctx;
-    };
-};
-
-template<typename T>
-class LargeIndirectContext {
-private:
-    Array<B16<T, 7>> data;
-    const uint32_t hashBits, inputBits;
-
-public:
-  LargeIndirectContext(const int hashBits, const int inputBits) :
-      data(UINT64_C(1) << hashBits),
-      hashBits(hashBits),
-      inputBits(inputBits)
-    {
-      assert(hashBits > 0 && hashBits <= 24);
-      assert(inputBits > 0 && inputBits <= 8);
-    }
-
-    void reset() {
-      for (uint64_t i = 0; i < data.size(); ++i) {
-        data[i].reset();
-      }
-    }
-
-    void set( uint64_t contextHash, const uint8_t c) {
-      contextHash=contextHash*987654323;
-      contextHash=contextHash<<16|contextHash>>16;
-      contextHash=contextHash*123456791;
-      assert(c < (1 << inputBits));
-      uint32_t* ptr = data[finalize64(contextHash, hashBits)].find(checksum16(contextHash, hashBits), nullptr);
-      *ptr = (*ptr) << inputBits | c;
-    };
-
-    uint32_t get(  uint64_t contextHash) {
-      contextHash=contextHash*987654323;
-      contextHash=contextHash<<16|contextHash>>16;
-      contextHash=contextHash*123456791;
-      return *data[finalize64(contextHash, hashBits)].find(checksum16(contextHash, hashBits), nullptr);
-    };
-
-};
 
 class MTFList{
 private:
