+++ prt/rnd.hpp
@@ -1,3 +1,4 @@
+#pragma once
 //////////////////////////// rnd ///////////////////////////////
 
 // 32-bit pseudo random number generator
